backend/user-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.0
python-multipart==0.0.9
